Split InputDatasets out of babs.py.
