+++ babs/cli.py
@@ -10,7 +10,8 @@ from pathlib import Path
 import pandas as pd
 from filelock import FileLock, Timeout
 
-from babs.babs import BABS, InputDatasets, System
+from babs.babs import BABS, System
+from babs.dataset import InputDatasets
 from babs.utils import (
     ToDict,
     _path_does_not_exist,
